Support author filter for Jetpack and self hosted
To check if a blog has multiple authors, we were already fetching a list of
users. Besides counting them, we can match the existing known username to
figure out the user ID, so we can use that for the "Only me" filter.

Previously, this was problematic for Jetpack sites, as the known user ID was a
wordpress.com user ID, which didn't match the site's user ID that the API was
using.

Also, this enables filtering for non-Jetpack/XML-RPC self hosted sites, since
there's no reason not to :)