Improve handling of redirects when adding sites
Most HTTP redirects will turn a POST request into a GET request, which makes
the XML-RPC probe fail. This means that a site with enforced SSL would require
multiple attempts before succeeding.

Now if we detect a redirect we'll retry a request. I've limited this to
redirects with the same host name.