Prevent triggering lazy property initialization
Attempt to fix crash (#14787) when WordPressOrgXMLRPCApi is being deinitialized.

The lazy vars are initialized from deinit, which is NOT what we want. If the vars have not been initialized already, we don't want to initialize them inside of deinit.

A question was asked about possibly adding a feature like this into Swift: https://forums.swift.org/t/lazy-var-and-deinit/2291