Add SwiftLint via CocoaPods
I find this is a simpler way to add SwiftLint because, by using a tool
that is already a required part of the setup, it doesn't need extra
scripts or commands that devs may forget to run.

In the long run, we'll want to install SwiftLint in a Ruby-less way.
This commit comes in the context of a HACK Week project, where I'm more
interested in bringing stuff to a working state soon rather than
crafting the best solution.

We can always iterate on top of this and change the SwiftLint install
mode later on.