Check error domain and code, instead of casting to WordPressComRestApiError
This is an unfortunate comprosie we need to make during a refactor(*).
After the refactor is done, the `error` variable here should be of type
`WordPressAPIError<WordPressComRestApiEndpointError>`. That means,
at the end, we can access the 'error code' in a more strongly-typed way.

(*): A WordPressComRestApi refactor to make it return
`WordPressAPIError` in its Swift API, instead of `NSError`.