Use `XCTUnwrap` instead of force-unwrapping for fixture paths in tests
This will make it easier to run the tests once the fixture files move
around. If a file is misplaced, the test will fail instead of crashing.
Operationally, failing is better than crashing because it generates a
list of all misplaced tests, so that they can be addressed in one go.
The alternative would be a run tests, tests crash, fix individual file
path loop for as many misplaced fixture files there are.