Remove all "Group without Folder" from WordPressKitTests
This is to prepare the move to `Sources/` and `Tests/`. Many of the
groups had only one file in them and there was no correspondence between
them and the source code. By removing them, the job of changing the root
to `Tests/` becomes simpler because we won't have to update references
to "Group without Folder" that could not update automatically.