Remove functionally-empty `ServiceRemoteRESTTests`
This test didn't exercise and specific behavior, just that the type
returned by the `init` was the object's type and that the `api`
parameter was set. We can do without those checks given the service
behavior is tested elsewhere via its subclasses and those settings are
therefore verified many many times.