Remove `PostServiceRemoteExtended` and implementations
They will be moved into WordPress and Jetpack.

The reason for this is that they are only used in WordPress and Jetpack
and that they are Swift extensions of Objective-C types which we are
trying to isolate.

It's simpler to move these leaf types into the only consumer that uses
them than trying to re-architect them.